feat: remove hero admin management
